removed unused components from the homepage. added that content to /app/page
