--- components/About/Description.js
@@ -1,24 +1,0 @@
-const About = () => {
-  return (
-    <div className="h-fit w-fit md:flex md:flex-col md:justify-center lg:flex lg:w-full">
-      <div className="space-y-2 text-base indent-4 text-neutral-300">
-        <span className="font-bold text-yellow-600">Enhance </span>
-        <span>
-          is a dynamic research group specializied in sports and esports
-          management.
-        </span>
-        <p>
-          Covering a diverse variety of topics our group brings together a team
-          of researchers with a wide range of expertise.
-        </p>
-        <p>
-          Our collaborative approach enables us to explore these themes
-          comprehensively and develop innovative solutions across various
-          domains within sports and esports management.
-        </p>
-      </div>
-    </div>
-  );
-};
-
-export default About;
