--- components/About/Header.js
@@ -1,14 +1,0 @@
-import { oxanium } from "../fonts";
-
-const Header = () => {
-  return (
-    <div
-      className={`${oxanium.className} text-3xl font-bold
-    uppercase tracking-wider text-yellow-600`}
-    >
-      <p>What is the Enhance Group?</p>
-    </div>
-  );
-};
-
-export default Header;
